requirements.txt
boto3>=1.34.0
pandas>=2.2.0
python-dotenv>=1.0.1
pytest>=8.2.0
Pillow>=10.0.0
requests>=2.31.0
